refactor: view 폴더 생성, 파일명 & 변수명 변경
