+++ src/js/constants/lottoConstraint.js
@@ -1,7 +1,8 @@
 export const PERCENTAGE_UNIT = 100;
 
+export const PRICE_UNIT = 1000;
+
 export const LOTTO_CONSTRAINT = {
-  PRICE_UNIT: 1000,
   MAX_IN_LOTTO_NUMBER: 45,
   MIN_IN_LOTTO_NUMBER: 1,
   LOTTO_NUMBERS_COUNT: 6,
